requirements.txt
google-genai
langgraph
langchain
langchain-community
chromadb
sentence-transformers
pypdf
pdfplumber
streamlit
python-dotenv
